Mannequin Core Files
These are the Mannequin Core Files that can be used to produce the Mannequin Core Structure developed by Shelly Uhlir and Adam Bradshaw. The PDF files have been laid out so that they can be printed on 8 1/2 X 11 paper and taped together to form a cutting template. The EPS and DXF files are Vector line drawings of the parts that can be used in any CNC output.